+++ packages/components/molecules/f-rating/test/visual/f-rating.visual.spec.js
@@ -107,23 +107,23 @@ devices.forEach(device => {
             });
         });
 
-        describe('`shouldAlignRatingLeft`', () => {
+        describe('`shouldAlignRatingTextLeft`', () => {
             it('should be displayed on the left hand side of the ratings with a default font size', async () => {
                 // Act
-                await Rating.load({ shouldAlignRatingLeft: true, reviewCount: 5, ratingDisplayType: 'short' });
+                await Rating.load({ shouldAlignRatingTextLeft: true, reviewCount: 5, ratingDisplayType: 'short' });
 
                 // Assert
-                await browser.percyScreenshot('f-rating Multi Star - Visual Test for Prop - shouldAlignRatingLeft = true', device);
+                await browser.percyScreenshot('f-rating Multi Star - Visual Test for Prop - shouldAlignRatingTextLeft = true', device);
             });
         });
 
         describe('`ratingFontSize`', () => {
             it('should be displayed on the left hand side of the ratings with a `large` font size', async () => {
                 // Act
-                await Rating.load({ shouldAlignRatingLeft: true, ratingFontSize: 'large', ratingDisplayType: 'short' });
+                await Rating.load({ shouldAlignRatingTextLeft: true, ratingFontSize: 'large', ratingDisplayType: 'short' });
 
                 // Assert
-                await browser.percyScreenshot('f-rating Multi Star - Visual Test for Prop - ratingFontSize = large & shouldAlignRatingLeft = true', device);
+                await browser.percyScreenshot('f-rating Multi Star - Visual Test for Prop - ratingFontSize = large & shouldAlignRatingTextLeft = true', device);
             });
         });
     });