+++ packages/components/molecules/f-rating/src/tests/constants.test.js
@@ -1,6 +1,7 @@
 import {
     VALID_STAR_RATING_SIZES,
-    VALID_STAR_RATING_DISPLAY_TYPE
+    VALID_STAR_RATING_DISPLAY_TYPE,
+    VALID_STAR_FONT_SIZES
 } from '../constants';
 
 describe('`constants`', () => {
@@ -19,11 +20,18 @@ describe('`constants`', () => {
             // Arrange
             const values = ['short', 'medium', 'long'];
 
-            // Act
-            const result = VALID_STAR_RATING_DISPLAY_TYPE.every(value => values.includes(value));
+            // Assert
+            expect(VALID_STAR_RATING_DISPLAY_TYPE).toEqual(values);
+        });
+    });
+
+    describe('`VALID_STAR_FONT_SIZES`', () => {
+        it('should contain the correct font types', () => {
+            // Arrange
+            const values = ['default', 'large'];
 
             // Assert
-            expect(result).toBe(true);
+            expect(VALID_STAR_FONT_SIZES).toEqual(values);
         });
     });
 });